Implement authorization bearer token user prompt
If there is no saved token, we show the Spotify web site and intercept the user's login to grab the token and store it. Then we continue as usual. If the token is stored, we grab it from the disk cache and get right to business.

This doesn't solve for token expiration, which I will implement later most probably by extracting the token prompt logic to its own file and re-displaying the Spotify web page window so the user can login again and refresh the token.