Switch back to CSS based rendering
This is better than the SVG because I can use the full palette of what CSS has to offer including auto-sizing the element containing the text.

I have also changed the way the lyrics are passed to a data attribute on the body and made the lyrics display a pseudo-element because it is the only thing that will ever be shown there so this feels a bit better than a dedicated element with an ID.

Doesn't solve the ghosting, still.