Clear up some warnings related to opening the Spotify website
Disabled CSP with `webSecurity` but it doesn't seem to do anything so I set `ELECTRON_DISABLE_SECURITY_WARNINGS`. If the former does work, good, if not, at least I don't care about the warnings.

The latter I couldn't figure out how to disable or just dismiss the warning so I enabled it to not get the warning.